I added the references it was missing. Compiling still did not work, due to "natbib" errors. I checked the \bibliography{} reference, the .aux file, tried multiple packages, etc.
It does not seem to be reading the .bib file. The only thing I could think of, is that I have extra sources in the .bib file than what I reference in the .rnw
The forums I looked at did not make this seem like a problem.

Therefore, I will return to this later.